Refactor the authentication endpoints
- Let's move the auth endpoints to the the new AuthController.
- Implement the browser-logout endpoint which clears the session.
