+++ build.sbt
@@ -337,7 +337,12 @@ lazy val server = (project in file("server"))
       "software.amazon.awssdk" % "ses" % "2.17.141",
       "jakarta.xml.bind" % "jakarta.xml.bind-api" % "3.0.1",
       "org.apache.commons" % "commons-text" % "1.9",
-      "io.swagger" % "swagger-annotations" % swagger
+      "io.swagger" % "swagger-annotations" % swagger,
+      // JAX-B dependencies for JDK 9+, required to use play sessions
+      "javax.xml.bind" % "jaxb-api" % "2.3.1",
+      "javax.annotation" % "javax.annotation-api" % "1.3.2",
+      "javax.el" % "javax.el-api" % "3.0.0",
+      "org.glassfish" % "javax.el" % "3.0.0"
     )
   )
