+++ lib/api/shared/src/main/scala/net/wiringbits/api/ApiClient.scala
@@ -13,6 +13,7 @@ trait ApiClient {
   def createUser(request: CreateUser.Request): Future[CreateUser.Response]
   def login(request: Login.Request): Future[Login.Response]
   def loginBrowser(request: Login.Request): Future[Login.Response]
+  def logoutBrowser(): Future[Logout.Response]
 
   def verifyEmail(request: VerifyEmail.Request): Future[VerifyEmail.Response]
   def forgotPassword(request: ForgotPassword.Request): Future[ForgotPassword.Response]
@@ -135,7 +136,7 @@ object ApiClient {
     }
 
     override def login(request: Login.Request): Future[Login.Response] = {
-      val path = ServerAPI.path :+ "users" :+ "login"
+      val path = ServerAPI.path :+ "auth" :+ "login"
       val uri = ServerAPI.withPath(path)
 
       prepareRequest[Login.Response]
@@ -147,7 +148,7 @@ object ApiClient {
     }
 
     override def loginBrowser(request: Login.Request): Future[Login.Response] = {
-      val path = ServerAPI.path :+ "users" :+ "login" :+ "browser"
+      val path = ServerAPI.path :+ "auth" :+ "login-browser"
       val uri = ServerAPI.withPath(path)
 
       prepareRequest[Login.Response]
@@ -158,8 +159,20 @@ object ApiClient {
         .flatMap(Future.fromTry)
     }
 
+    override def logoutBrowser(): Future[Logout.Response] = {
+      val path = ServerAPI.path :+ "auth" :+ "logout-browser"
+      val uri = ServerAPI.withPath(path)
+
+      prepareRequest[Logout.Response]
+        .post(uri)
+        .body(Json.toJson(Logout.Request()).toString())
+        .send(backend)
+        .map(_.body)
+        .flatMap(Future.fromTry)
+    }
+
     override def currentUser(jwt: String): Future[GetCurrentUser.Response] = {
-      val path = ServerAPI.path :+ "users" :+ "me"
+      val path = ServerAPI.path :+ "auth" :+ "me"
       val uri = ServerAPI.withPath(path)
 
       prepareRequest[GetCurrentUser.Response]
